Implementing image cell component
