+++ stories/gallery/imageCell.stories.js
@@ -1,0 +1,24 @@
+import ImageCell from '../../assets/cms/components/gallery/ImageCell'
+
+export default {
+    title: 'Gallery/Image Cell',
+    component: ImageCell
+}
+
+export const basicUsage = () => ({
+    components: {
+        ImageCell,
+        ...ImageCell.components
+    },
+    data: () => ({
+        clicks: 0,
+        clickDelete: 0
+    }),
+    template: `
+    <div class="ccm-ui">
+        <p>Clicks: {{clicks}}</p>
+        <p>Click Delete: {{clickDelete}}</p>
+        <ImageCell src="https://via.placeholder.com/120" file-size="104 KB" size="120" @click="clicks++" @delete="clickDelete++"/>
+    </div>
+    `
+})
